Project added
This is the project in the state that I found it. I realize I never set it up with any sort of Source control. My first move is going to be to scale back some of the accomplishments and focus on improving the code quality of the basics. This version of the project will be slowly rebuilt but this time with the correct ideals in mind.